Split settings preload bundles out of features/index.ts; fix types/lint
